refactor: metrics and applications apis and utils (#98)
* chore: refactored src/utils/fetchAttestations.ts

separated concerns in different files by functions and types.
NOTE: this change does not affect the codebase as everything that was exported by fetchAttestations.ts now is exported by src/utils/fetchAttestations/index.ts

* chore: refactored api/routers/applications

separated concerns in different files extracting the logic of fetching applications and approvals from attestations to be reused by other routers in the server side
NOTE: doesn't affect the codebase as everything exported previously by applications.ts is exported now by applications/index.ts

* chore: refactored src/utils/fetchMetrics

separated concerns in different files by functions and types.
NOTE: this change does not affect the codebase as everything that was exported by fetchMetrics.ts now is exported by src/utils/fetchMetrics/index.ts

* chore: add missing return type to createAttestationFetcher()

* chore: created trpc ballotAttestationProcedure

added type keyword to some imports

* chore: refactored api/routers/metrics

separated concerns in different files extracting types, mocks and getter functions
NOTE: doesn't affect the codebase as everything exported previously by metrics.ts is exported now by metrics/index.ts

* chore: implemented fetchApprovedApplications
