+++ src/server/api/routers/metrics/index.ts
@@ -1,0 +1,1 @@
+export { metricsRouter } from "./metricsRouter";
