+++ src/server/api/routers/metrics/mocks.ts
@@ -1,0 +1,7 @@
+export const mockedApprovedProjects = [
+  "zora",
+  "uniswap",
+  "aave",
+  "gitcoin",
+  "layer3xyz",
+];
