+++ src/utils/fetchMetrics/mocks.ts
@@ -1,0 +1,2 @@
+export const mockedDescription =
+  "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Sed et ligula at dolor bibendum auctor. Fusce eu justo nec nisl suscipit sollicitudin. Aliquam erat volutpat. Mauris quis facilisis purus, non auctor magna. Nulla facilisi. Donec et odio vel nulla ornare viverra. Vivamus at lacus nec urna sodales fermentum. Pellentesque habitant morbi tristique senectus et netus et malesuada fames ac turpis egestas. Cras sagittis, nibh in sodales aliquet, libero mi porttitor elit, non pharetra arcu nulla non est. Proin vitae augue dignissim, malesuada enim non, fermentum nisi.";
